analysis of  xray images using auto encoders
Autoencoders are neural networks that you can use to compress and reconstruct data. The encoder compresses input, and the decoder attempts to recreate the information from this compressed version. Autoencoders are a deep learning model for transforming data from a high-dimensional space to a lower-dimensional space. They work by encoding the data, whatever its size, to a 1-D vector. This vector can then be decoded to reconstruct the original data (in this case, an image).